Added lookup/autocreation of PI Points
Tags can now be added without knowing their WebID.  WebIDs are looked
up by their tagname and if the PI Point does not exist on the server
it is created.

There is a limitation of the RequestHTTPX method that does not allow
access of the response header.  This forces a Get request after the
PI Point is created to access it's WebID rather than just reading its
WebID from the response headers when it is created.