fix type creation format field bug
OMF has an optional format field to use non default types, They dont exist
for strings and bools but they do for the others so this was backwards but
it was still working for whatever reason. it was probably working because
there was a missing comma at the end making it invalid json so it seems
to not update.