Improve memory handling to prevent OOM crashes
Removes the explicit sizing of StringBuffers to
prevent the allocation of memory which may not
be used.

The allocation of extra, unused memory will
significantly reduce the available heap size and
may starve other applications or cause the JVM to
crash. This is especially applicable when using
this connector with other applications that make
use of the Java memory heap.

Additionally, a memory check was added to queue
data polls to ensure that a reasonable amount of
memory is free before continuing.